+++ app/src/main/java/com/vsantander/blockchaininfochallenge/domain/model/TransactionPerSecond.kt
@@ -1,6 +1,6 @@
 package com.vsantander.blockchaininfochallenge.domain.model
 
 data class TransactionPerSecond(
-        val timestamp: Double,
-        val numberBitcoins: Double
+        val timestamp: Float,
+        val numberBitcoins: Float
 )
